Rejig some functions and move a file
